Fix token-swap tests
Token-swap tests currently do not pass, even with the linking
issue resolved on `create_program_address`, failing with key errors
since there are incorrect hard-coded program ids.

To fix this, some more conditional compilation is done to create valid
program ids for token and token-swap, followed by properly adding
the program as a signer on its addresses, mimicking what happens in the
real runtime's syscalls.

Note: In order to get valid `Keypair`s as they are used by the BPF
loader, the `program` feature is removed from the test build, since it
requires `sdk::signature::Keypair`, which is not allowed for
`feature = program`.  I considered adding the ed25519-dalek library as
a dev dependency, but it seemed better to use the same code as the sdk.
Let me know if there's a better way to generate these while also
maintaining `feature = program`.