fix ConcurrentMerkleTreeHeader (de)serialization
The `account_type` member of `ConcurrentMerkleTreeHeader` is currently
an `#[repr(u32)]` enum, but borsh only uses one byte to serialize or
deserialize an enum tag, which leads to issues with `try_from_slice`
calls that get passed slices computed based on applying `size_of` to
`ConcurrentMerkleTreeHeader`. This commit replaces the underlying
type of `account_type` with an integer type.

Signed-off-by: Alexandru Agache <alexandruag@gmail.com>