feat: Add compatibility for publicKeys
Ran into an issue where I was passing web3.publicKeys to createMintToInstruction but received an anonymous.toString() is undefined error. 

The error was that the docs allowed a (signer | publicKey) [ ] but the function only works with a signer[ ], my fix was to move my publicKeys into an object. {publicKey: myKey}

I believe the changes to the logic in this file will resolve the issue for anyone encountering it in the future