deps: Bump sha2 to 0.9.9
#### Problem

The binary built with `cargo install --locked spl-token-cli` has a
segfault, likely due to https://github.com/RustCrypto/hashes/issues/344.
The normal build with `cargo install spl-token-cli` does work, however.
It's not clear why this issue has only started now, but the
spl-token-cli build is not working on the Solana 2.0.4 and 2.0.5
releases.

#### Solution

Bump sha2 to 0.9.9.