CI: Add manual workflow to publish crates
#### Problem

There are a lot of crates in SPL, and in the future SDK monorepo, so it
would be extremely useful to have a manual GitHub Actions workflow to
test a package, bump its version everywhere, publish it, and create a
release on GitHub.

#### Summary of changes

Following the excellent scripts at the
[counter-shank repo](https://github.com/solana-program/counter-shank/blob/main/.github/workflows/publish-rust-client.yml),
port the zx bits to bash, and do mostly the same thing, except allowing
for multiple packages to be published. Thanks @lorisleiva!