CI: Add git-cliff step during publish
#### Problem

The publish step is great, but there's still some manual work to add the
changelog and correct title for the created release.

#### Summary of changes

Use git-cliff to generate the release notes. This comes with a very
simple cliff.toml which will do minimal processing, since we don't
follow conventional commits in the repo.

Here's a test run for the given input:

```
$ git-cliff -c ci/cliff.toml "pod-v0.3.0"..master --include-path "libraries/pod/**" --github-repo "solana-labs/solana-program-library"
```

Output:

```
## What's new

- Publish pod v0.3.2 by @github-actions[bot]
- [token-2022] Upgrade to `zk-sdk` (#7148) by @samkim-crypto
- Implement `Default` for `PodOption` (#7083) by @joncinque
- Bump to 0.3.1 (#7075) by @febo
- Improve `PodOption` type (#7076) by @febo
- Add `PodU128` type (#7012) by @febo
- Add `PodOption` type (#6886) by @febo
- Use `bytemuck_derive` explicitly (#6928) by @joncinque
```