ci: Rebase during publish step
#### Problem

Sometimes, while the publish workflow is running, we merge a different
PR, which means that the workflow is running against an older commit.
This makes the step to push the tag commit fail sometimes, but this
happens *after* the new crate has already been published. This means
that we end up publishing crates that don't have any commits attached to
them.

#### Summary of changes

Add a step to rebase before publishing.