ci: Add git tag / commit message with cargo release
#### Problem

The publish workflow is great, but it does some work by hand, namely
creating the git commit and tag. `cargo release` can do this, but we
currently don't use it because we aren't customizing it.

#### Summary of changes

Add configuration options to the top-level Cargo.toml and let cargo
release do everything for us.